allow prefix matching for actions
This offers some command line convenience, allowing the user only
specify an action prefix as long as it can be resolved without
ambiguitiy. For example, as of this commit, 'asp up' would then
translate to 'asp update', and 'asp d' would be an error (as 'difflog'
and 'disk-usage' are both candidates).