hide git repo prefix in list-{all,local} output
This is an implementation detail -- no one should really need to know
or core which repo the package came from. Moreover, there's nothing you
can do with this information. It's conceivable that a package of the
same name could exist in both git repos, but this would be semantically
forbidden by Arch packaging policies, and not well handled by asp
regardless of showing the prefix here.

Fixes GH#14.