export: handle arch=any packages when a repo is specified
Package might be arch=any, which means that they don't have the
architecture-specific branches. Duh. But 'any' also isn't an
architecture. It's unfair to make people understand this. Rather, asp
should assume that when someone writes something like:

  $ asp export extra/asp

They want the snapshot of 'asp' from the 'extra' repo that's appropriate
for their architecture. In this case, that's any, and not $(uname -m).

Fixes #32.