update: recreate local branches if they do not exist
If the asp cache gets purged for some reason, local checkouts of packages will
continue to refer to (now inexistent) local branches of the cache git repo. The
only way to recreate those branches is to repeat an `asp checkout` of each
affected package, but there is no reason to do that.

Make sure that a simple `asp update` is sufficient to recreate those branches.

Fixes #35.